20230125 hozin2/1

+++ src/main/java/com/deliverygig/moonjyoung/service/FoodCategoryService.java
@@ -34,6 +34,7 @@ public class FoodCategoryService {
             resultMap.put("code", HttpStatus.BAD_REQUEST);
             return resultMap;
         }
+        else {
         FoodCategoryEntity fcEntity = FoodCategoryEntity.builder()
             .fcName(data.getFcName())
             .storeInfoEntity(siRepo.findBySiSeq(data.getFcSiSeq())) //.fcSiSeq(data.getFcSiSeq())
@@ -45,6 +46,7 @@ public class FoodCategoryService {
         resultMap.put("message", "등록이 완료되었습니다.");
         resultMap.put("code", HttpStatus.ACCEPTED);
         return resultMap;
+        }
     }
 
     // read
@@ -68,20 +70,15 @@ public class FoodCategoryService {
         return resultMap;
     }
     // update
-    public Map<String, Object> cateUpdate(String seq, String value, CateUpdateVO data) {
+    public Map<String, Object> cateUpdate(Long seq, CateUpdateVO data) {
         Map<String, Object> resultMap = new LinkedHashMap<String, Object>();
-        if((long)Integer.parseInt(value) != data.getFcSeq()) {
+        if(fcRepo.findByFcSeq(seq)==null) {
             resultMap.put("status", false);
             resultMap.put("code", HttpStatus.BAD_REQUEST);
-            resultMap.put("message", "seq번호를 일치시키세요.");
+            resultMap.put("message", "seq번호를 확인하세요.");
             return resultMap;
         }
         else {
-            FoodCategoryEntity entity = fcRepo.findById(data.getFcSeq()).get();
-                entity.setFcName(data.getFcName());
-                entity.setStoreInfoEntity(siRepo.findBySiSeq(data.getFcSiSeq()));
-                entity.setFcOrder(data.getFcOrder());
-
             if(fcRepo.countByFcName(data.getFcName()) == 1) {
                 resultMap.put("status", false);
                 resultMap.put("message", data.getFcName() + "은/는 이미 존재하는 카테고리명 입니다.");
@@ -89,6 +86,11 @@ public class FoodCategoryService {
                 return resultMap;
             }
             else{
+                FoodCategoryEntity entity = fcRepo.findByFcSeq(seq);
+                entity.setFcName(data.getFcName());
+                entity.setStoreInfoEntity(siRepo.findBySiSeq(data.getFcSiSeq()));
+                entity.setFcOrder(data.getFcOrder());
+                
                 fcRepo.save(entity);
                 resultMap.put("status", true);
                 resultMap.put("message", "수정이 완료되었습니다.");
@@ -102,11 +104,16 @@ public class FoodCategoryService {
     @Transactional
     public Map<String, Object> dCate(@RequestBody FoodCategoryEntity data) {
         Map<String, Object> resultMap = new LinkedHashMap<String, Object>();
-
-        fcRepo.deleteByFcSeq(data.getFcSeq());
-        resultMap.put("status", true);
-        resultMap.put("message", "삭제 되었습니다.");
-        resultMap.put("code", HttpStatus.OK);
+        if(fcRepo.countByFcSeq(data.getFcSeq()) == 1) {
+            fcRepo.deleteByFcSeq(data.getFcSeq());
+            resultMap.put("status", true);
+            resultMap.put("message", "삭제 되었습니다.");
+            resultMap.put("code", HttpStatus.OK);
+            return resultMap;
+        }
+        resultMap.put("status", false);
+        resultMap.put("message", "카테고리 Seq 오류입니다.");
+        resultMap.put("code", HttpStatus.BAD_REQUEST);
         return resultMap;
     }
